Add files via upload
This is an updated version of the previous iteration. This expands on harvesting sc2 matchmaking loading screen information by detecting the opponent's Match Making Rating, Race (terran, zerg, protoss) and in-game name. It outputs to a txt file so that it can be easily viewed on a twitch stream through and OBS read from txt file graphic.